requirements.txt
Flask==3.1.1
pdf2image==1.17.0
Pillow==11.3.0
pywin32==305.1
waitress==3.0.2
pypdf==5.8.0